backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.21.0
werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2